Prevent crashes when trying to get null data
Gozk crashes when a Java-based client create a null zknode; e.g. `create
/mynode` using zkCli.sh.  This patch, sourced from
https://bugs.launchpad.net/gozk/+bug/1302789, fixes this problem.
From the description there, zoo_wget returns -1 in buffer_len, which is
interpreted as a very large string size, which leads to OOM.  This patch
returns an empty string if the znode is zero size, and the actual data
otherwise.