Documenting build process
Building this library started breaking when we introduced the M1
laptops. Initially it appeared to be an issue with the new instruction
set, however, builds failed on linux as well. The main issue was that
the build procedure was not properly documented.

Due to that, header files and shared libraries could not be found, and
even after rectifying this, a preprocessor define (`THREADED`) was
missing the compiler couldn't resolve all the symbols.

This commit updates the README with necessary instructions on how to
build, and adds CGO flags specifically for building on the new M1
macbooks.  Further it adds a `go.mod` file to properly declare
dependencies.